Move Config to base and unsafe
