+++ oak_functions/loader/Cargo.toml
@@ -40,8 +40,6 @@ oak_functions_abi = { path = "../abi" }
 oak_functions_extension = { path = "../extension" }
 oak_functions_lookup = { path = "../lookup" }
 oak_functions_util = { path = "../util" }
-oak_functions_metrics = { path = "../experimental/metrics" }
-oak_functions_tf_inference = { path = "../experimental/tf_inference" }
 oak_functions_wasm = { path = "../wasm" }
 oak_functions_workload_logging = { path = "../workload_logging" }
 oak_logger = { path = "../logger" }
